Refactor CupertinoScaffold
