+++ packages/flutter/lib/cupertino.dart
@@ -13,10 +13,11 @@ export 'src/cupertino/button.dart';
 export 'src/cupertino/colors.dart';
 export 'src/cupertino/dialog.dart';
 export 'src/cupertino/nav_bar.dart';
-export 'src/cupertino/page.dart';
-export 'src/cupertino/scaffold.dart';
+export 'src/cupertino/page_scaffold.dart';
+export 'src/cupertino/route.dart';
 export 'src/cupertino/slider.dart';
 export 'src/cupertino/switch.dart';
+export 'src/cupertino/tab_scaffold.dart';
 export 'src/cupertino/text_selection.dart';
 export 'src/cupertino/thumb_painter.dart';
 export 'widgets.dart';
