+++ packages/flutter/lib/src/cupertino/tab_scaffold.dart
@@ -1,0 +1,154 @@
+// Copyright 2017 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/widgets.dart';
+import 'bottom_tab_bar.dart';
+
+/// Implements a tabbed iOS application's root layout and behavior structure.
+///
+/// The scaffold lays out the tab bar at the bottom and the content between or
+/// behind the tab bar.
+///
+/// A [tabBar] and a [rootTabPageBuilder] are required. The [CupertinoTabScaffold]
+/// will automatically listen to the provided [CupertinoTabBar]'s tap callbacks
+/// to change the active tab.
+///
+/// Tabs' contents are built with the provided [rootTabPageBuilder] at the active
+/// tab index. [rootTabPageBuilder] must be able to build the same number of
+/// pages as the [tabBar.items.length]. Inactive tabs will be moved [Offstage]
+/// and its animations disabled.
+// TODO(xster): describe navigator handlings.
+// TODO(xster): add an example.
+class CupertinoTabScaffold extends StatefulWidget {
+  const CupertinoTabScaffold({
+    Key key,
+    @required this.tabBar,
+    @required this.rootTabPageBuilder,
+  }) : assert(tabBar != null),
+       assert(rootTabPageBuilder != null),
+       super(key: key);
+
+  /// The [tabBar] is a [CupertinoTabBar] drawn at the bottom of the screen
+  /// that lets the user switch between different tabs in the main content area
+  /// when present.
+  ///
+  /// When provided, [CupertinoTabBar.currentIndex] will be ignored and will
+  /// be managed by the [CupertinoTabScaffold] to show the currently selected page
+  /// as the active item index. If [CupertinoTabBar.onTap] is provided, it will
+  /// still be called. [CupertinoTabScaffold] automatically also listen to the
+  /// [CupertinoTabBar]'s `onTap` to change the [CupertinoTabBar]'s `currentIndex`
+  /// and change the actively displayed tab in [CupertinoTabScaffold]'s own
+  /// main content area.
+  ///
+  /// If translucent, the main content may slide behind it.
+  /// Otherwise, the main content's bottom margin will be offset by its height.
+  final CupertinoTabBar tabBar;
+
+  /// An [IndexedWidgetBuilder] that's called when tabs become active.
+  ///
+  /// When the tab becomes inactive, its content is still cached in the widget
+  /// tree [Offstage] and its animations disabled.
+  ///
+  /// Content can slide under the [tabBar] when it's translucent.
+  final IndexedWidgetBuilder rootTabPageBuilder;
+
+  @override
+  _CupertinoTabScaffoldState createState() => new _CupertinoTabScaffoldState();
+}
+
+class _CupertinoTabScaffoldState extends State<CupertinoTabScaffold> {
+  int _currentPage = 0;
+
+  @override
+  Widget build(BuildContext context) {
+    final List<Widget> stacked = <Widget>[];
+
+    // The main content being at the bottom is added to the stack first.
+    stacked.add(
+      new Padding(
+        padding: new EdgeInsets.only(bottom: widget.tabBar.opaque ? 0.0 : widget.tabBar.preferredSize.height),
+        child: new _TabView(
+          currentTabIndex: _currentPage,
+          tabNumber: widget.tabBar.items.length,
+          rootTabPageBuilder: widget.rootTabPageBuilder,
+        )
+      ),
+    );
+
+    if (widget.tabBar != null) {
+      stacked.add(new Align(
+        alignment: FractionalOffset.bottomCenter,
+        // Override the tab bar's currentIndex to the current tab and hook in
+        // our own listener to update the _currentPage on top of a possibly user
+        // provided callback.
+        child: widget.tabBar.copyWith(
+            currentIndex: _currentPage,
+            onTap: (int newIndex) {
+              setState(() {
+                _currentPage = newIndex;
+              });
+              // Chain the user's original callback.
+              if (widget.tabBar.onTap != null)
+                widget.tabBar.onTap(newIndex);
+            }
+        ),
+      ));
+    }
+
+    return new Stack(
+      children: stacked,
+    );
+  }
+}
+
+/// An widget laying out multiple tabs with only one active tab being built
+/// at a time and on stage. Off stage tabs' animations are stopped.
+class _TabView extends StatefulWidget {
+  _TabView({
+    @required this.currentTabIndex,
+    @required this.tabNumber,
+    @required this.rootTabPageBuilder,
+  }) : assert(currentTabIndex != null),
+       assert(tabNumber != null && tabNumber > 0),
+       assert(rootTabPageBuilder != null);
+
+  final int currentTabIndex;
+  final int tabNumber;
+  final IndexedWidgetBuilder rootTabPageBuilder;
+
+  @override
+  _TabViewState createState() => new _TabViewState();
+}
+
+class _TabViewState extends State<_TabView> {
+  List<Widget> tabs;
+
+  @override
+  void initState() {
+    super.initState();
+    tabs = new List<Widget>(widget.tabNumber);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return new Stack(
+      children: new List<Widget>.generate(widget.tabNumber, (int index) {
+        final bool active = index == widget.currentTabIndex;
+
+        // TODO(xster): lazily replace empty tabs with Navigators instead.
+        if (active || tabs[index] != null)
+          tabs[index] = widget.rootTabPageBuilder(context, index);
+
+        return new Offstage(
+          offstage: !active,
+          child: new TickerMode(
+            enabled: active,
+            child: tabs[index] ?? new Container(),
+          ),
+        );
+      }),
+    );
+  }
+}
