+++ packages/flutter/test/cupertino/scaffold_test.dart
@@ -5,16 +5,9 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../rendering/rendering_tester.dart';
 import '../services/mocks_for_image_cache.dart';
 
-List<int> selectedTabs;
-
 void main() {
-  setUp(() {
-    selectedTabs = <int>[];
-  });
-
   testWidgets('Contents are behind translucent bar', (WidgetTester tester) async {
     await tester.pumpWidget(
       new WidgetsApp(
@@ -23,7 +16,7 @@ void main() {
           return new CupertinoPageRoute<Null>(
             settings: settings,
             builder: (BuildContext context) {
-              return const CupertinoScaffold(
+              return const CupertinoPageScaffold(
                 // Default nav bar is translucent.
                 navigationBar: const CupertinoNavigationBar(
                   middle: const Text('Title'),
@@ -49,14 +42,29 @@ void main() {
           return new CupertinoPageRoute<Null>(
             settings: settings,
             builder: (BuildContext context) {
-              return new CupertinoScaffold.tabbed(
-                navigationBar: const CupertinoNavigationBar(
-                  backgroundColor: CupertinoColors.white,
-                  middle: const Text('Title'),
+              return new CupertinoTabScaffold(
+                tabBar: new CupertinoTabBar(
+                  items: <BottomNavigationBarItem>[
+                    const BottomNavigationBarItem(
+                      icon: const ImageIcon(const TestImageProvider(24, 24)),
+                      title: const Text('Tab 1'),
+                    ),
+                    const BottomNavigationBarItem(
+                      icon: const ImageIcon(const TestImageProvider(24, 24)),
+                      title: const Text('Tab 2'),
+                    ),
+                  ],
                 ),
-                tabBar: _buildTabBar(),
                 rootTabPageBuilder: (BuildContext context, int index) {
-                   return index == 0 ? page1Center : new Stack();
+                  return index == 0
+                      ? new CupertinoPageScaffold(
+                        navigationBar: const CupertinoNavigationBar(
+                          backgroundColor: CupertinoColors.white,
+                          middle: const Text('Title'),
+                        ),
+                        child: page1Center,
+                      )
+                      : new Stack();
                 }
               );
             },
@@ -67,132 +75,4 @@ void main() {
 
     expect(tester.getSize(find.byWidget(page1Center)).height, 600.0 - 44.0 - 50.0);
   });
-
-  testWidgets('Tab switching', (WidgetTester tester) async {
-    final List<int> tabsPainted = <int>[];
-
-    await tester.pumpWidget(
-      new WidgetsApp(
-        color: const Color(0xFFFFFFFF),
-        onGenerateRoute: (RouteSettings settings) {
-          return new CupertinoPageRoute<Null>(
-            settings: settings,
-            builder: (BuildContext context) {
-              return new CupertinoScaffold.tabbed(
-                navigationBar: const CupertinoNavigationBar(
-                  backgroundColor: CupertinoColors.white,
-                  middle: const Text('Title'),
-                ),
-                tabBar: _buildTabBar(),
-                rootTabPageBuilder: (BuildContext context, int index) {
-                  return new CustomPaint(
-                    child: new Text('Page ${index + 1}'),
-                    painter: new TestCallbackPainter(
-                      onPaint: () { tabsPainted.add(index); }
-                    )
-                  );
-                }
-              );
-            },
-          );
-        },
-      ),
-    );
-
-    expect(tabsPainted, <int>[0]);
-    RichText tab1 = tester.widget(find.descendant(
-      of: find.text('Tab 1'),
-      matching: find.byType(RichText),
-    ));
-    expect(tab1.text.style.color, CupertinoColors.activeBlue);
-    RichText tab2 = tester.widget(find.descendant(
-      of: find.text('Tab 2'),
-      matching: find.byType(RichText),
-    ));
-    expect(tab2.text.style.color, CupertinoColors.inactiveGray);
-
-    await tester.tap(find.text('Tab 2'));
-    await tester.pump();
-
-    expect(tabsPainted, <int>[0, 1]);
-    tab1 = tester.widget(find.descendant(
-      of: find.text('Tab 1'),
-      matching: find.byType(RichText),
-    ));
-    expect(tab1.text.style.color, CupertinoColors.inactiveGray);
-    tab2 = tester.widget(find.descendant(
-      of: find.text('Tab 2'),
-      matching: find.byType(RichText),
-    ));
-    expect(tab2.text.style.color, CupertinoColors.activeBlue);
-
-    await tester.tap(find.text('Tab 1'));
-    await tester.pump();
-
-    expect(tabsPainted, <int>[0, 1, 0]);
-  });
-
-  testWidgets('Tabs are lazy built and moved offstage when inactive', (WidgetTester tester) async {
-    final List<int> tabsBuilt = <int>[];
-
-    await tester.pumpWidget(
-      new WidgetsApp(
-        color: const Color(0xFFFFFFFF),
-        onGenerateRoute: (RouteSettings settings) {
-          return new CupertinoPageRoute<Null>(
-            settings: settings,
-            builder: (BuildContext context) {
-              return new CupertinoScaffold.tabbed(
-                  navigationBar: const CupertinoNavigationBar(
-                    backgroundColor: CupertinoColors.white,
-                    middle: const Text('Title'),
-                  ),
-                  tabBar: _buildTabBar(),
-                  rootTabPageBuilder: (BuildContext context, int index) {
-                    tabsBuilt.add(index);
-                    return new Text('Page ${index + 1}');
-                  }
-              );
-            },
-          );
-        },
-      ),
-    );
-
-    expect(tabsBuilt, <int>[0]);
-    expect(find.text('Page 1'), findsOneWidget);
-    expect(find.text('Page 2'), findsNothing);
-
-    await tester.tap(find.text('Tab 2'));
-    await tester.pump();
-
-    // Both tabs are built but only one is onstage.
-    expect(tabsBuilt, <int>[0, 0, 1]);
-    expect(find.text('Page 1', skipOffstage: false), isOffstage);
-    expect(find.text('Page 2'), findsOneWidget);
-
-    await tester.tap(find.text('Tab 1'));
-    await tester.pump();
-
-    expect(tabsBuilt, <int>[0, 0, 1, 0, 1]);
-    expect(find.text('Page 1'), findsOneWidget);
-    expect(find.text('Page 2', skipOffstage: false), isOffstage);
-  });
-}
-
-CupertinoTabBar _buildTabBar() {
-  return new CupertinoTabBar(
-    items: <BottomNavigationBarItem>[
-      const BottomNavigationBarItem(
-        icon: const ImageIcon(const TestImageProvider(24, 24)),
-        title: const Text('Tab 1'),
-      ),
-      const BottomNavigationBarItem(
-        icon: const ImageIcon(const TestImageProvider(24, 24)),
-        title: const Text('Tab 2'),
-      ),
-    ],
-    backgroundColor: CupertinoColors.white,
-    onTap: (int newTab) => selectedTabs.add(newTab),
-  );
 }
